+++ test/fixtures/another_dependency_node.rb
@@ -1,2 +1,2 @@
-require 'dependency'
+require './dependency'
 require 'inside_load_path'