+++ test/fixtures/entry_arity.js
@@ -1,0 +1,1 @@
+require('./arity_1.rb')