+++ test/fixtures/js_require.rb
@@ -1,0 +1,2 @@
+require 'pure_js'
+require 'pure_js2.js'