+++ test/fixtures/load_path/inside_load_path.rb
@@ -1,0 +1,2 @@
+INSIDE=789
+puts 'we made it'