+++ test/fixtures/relative.rb
@@ -1,0 +1,2 @@
+# sprockets allows you to not lead with a '.'
+require_relative 'tree/file1'