+++ test/fixtures/tree/file1.rb
@@ -1,0 +1,2 @@
+HELLO=123
+puts 'inside the tree'